app/requirements.txt

Flask-MySQL
flask-mysqldb
pymysql
SQLAlchemy
Flask-SQLAlchemy
WTForms
flask
flask-wtf
seaborn
mysql-connector-python>8 
pandas>1
certifi 
cffi
netmiko
bokeh==2.4.2 




